requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
altgraph==0.17.5
anyio==4.13.0
attrs==25.3.0
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.7
et_xmlfile==2.0.0
fire==0.7.1
fonttools==4.62.1
frozenlist==1.8.0
h11==0.16.0
heif-convert==1.2.1
httpcore==1.0.9
httpx==0.28.1
idna==3.10
lxml==5.3.2
multidict==6.7.1
numpy==2.2.4
opencv-python-headless==4.13.0.92
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==26.0
pandas==2.2.3
pdf2docx==0.5.12
pefile==2024.8.26
pillow==12.2.0
pillow_heif==1.3.0
propcache==0.4.1
pycparser==2.22
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.4
PyMuPDF==1.27.2.2
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.2
python-telegram-bot==22.7
pytz==2025.2
pywin32-ctypes==0.2.3
requests==2.33.1
selenium==4.31.0
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
sortedcontainers==2.4.0
soupsieve==2.6
termcolor==3.3.0
tkinterdnd2==0.4.3
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.23.0
